requirements.txt
anyio==3.6.2
appdirs==1.4.4
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cryptography==39.0.0
cycler==0.11.0
fastapi==0.88.0
fonttools==4.38.0
frozendict==2.3.4
h11==0.14.0
html5lib==1.1
httptools==0.5.0
idna==3.4
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.6.2
mpl-finance==0.10.1
mplfinance==0.12.9b7
multitasking==0.0.11
nsepy==0.8
numpy==1.24.1
packaging==23.0
pandas==1.5.2
pandas-datareader==0.10.0
Pillow==9.4.0
plotly==5.11.0
pycparser==2.21
pydantic==1.10.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
tenacity==8.1.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
yfinance==0.2.3
